payment-aggregator/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1

